src/util/StringUtils: Split functions up into a header and cpp file.
