+++ src/parser/sparqlParser/CMakeLists.txt
@@ -4,7 +4,7 @@ add_library(sparqlParser SparqlQleverVisitor.h
         generated/SparqlAutomaticListener.h generated/SparqlAutomaticListener.cpp
         generated/SparqlAutomaticParser.h generated/SparqlAutomaticParser.cpp
         generated/SparqlAutomaticVisitor.h generated/SparqlAutomaticVisitor.cpp)
-qlever_target_link_libraries(sparqlParser antlr4_static sparqlExpressions rdfEscaping)
+qlever_target_link_libraries(sparqlParser antlr4_static sparqlExpressions rdfEscaping util)
 # Silence warnings in files that are auto-generated by ANTLR.
 # TODO<joka921> Submit a pull request to ANTLR to fix those warnings.
 target_compile_options(sparqlParser PRIVATE -Wno-logical-op-parentheses -Wno-parentheses)
